build fix: ath_hal_getrtsaggrlimit can be removed here
reason for this part of code:
"AR5416/AR5418 doesn't work with RTS + aggregates longer than 8KiB,
so the driver has to either limit aggregate sizes to 8KiB or less,
or disable RTS on longer aggregates. This was fixed in subsequent chips.
...
 it's quite possible that there's some carrier / enterprise solution
somewhere that uses the AR7010 as an offload CPU for an AR5418 device.
...
I'd eventually like us to support the ethernet offload mode -
even if it's just as a reference we can use internally
at work when demo'ing stuff."
  Adrian Chadd

Signed-off-by: Oleksij Rempel <linux@rempel-privat.de>