CMakeLists.txt: add -fno-lto
this option is needed to make compile work with cross-toolchain properly.
We are not using LTO any way, but for some reasons if cross-tolachain
is compailed without LTO, then FW compilation will fail.

With and without this option FW has same md5sum - So it should be save
to use this option.

Signed-off-by: Oleksij Rempel <linux@rempel-privat.de>