Add AppStream metadata file.
A slightly different one was kept downstream in Debian, but
this should be useful to other distros to, amongst other
things, help users install the missing package upon insertion
of the hardware.

The AppStream specification only permits certain permissive
licenses, because usually all packages' metadata needs to be
aggregated together by distros. The FSF All-Permissive license
is used, as BSD-3-Clause(-Clear) is not permitted by the
specification.