Remove last instance of BSD-4 advertisement clause
The Regents of the University of California allowed their BSD-4-clause
licensed code to delete the 3rd clause via a copyright addendum which is
described at https://www.freebsd.org/copyright/license.

For the file asf_queue.h, the 3rd clause was already removed, so remove it
in if_llc.h as well.

Signed-off-by: Bastian Germann <bage@debian.org>